requirements.txt
selene==2.0.0rc9
pytest==8.3.4
allure-pytest==2.13.5
selenium==4.28.1